fix clippy::unnecessary_cast

+++ rusvid_lib/src/effect/library/box_blur.rs
@@ -75,7 +75,7 @@ impl EffectLogic for BoxBlur {
                 let x = x as i32;
                 let y = y as i32;
 
-                let count = (self.kernel_x * self.kernel_y) as u32;
+                let count = self.kernel_x * self.kernel_y;
                 let sum = ((self.abs_d_y * -1)..=self.abs_d_y)
                     .cartesian_product((self.abs_d_x * -1)..=self.abs_d_x)
                     .map(|(i_x, i_y)| *original.pixel_unchecked((x + i_x) as u32, (y + i_y) as u32))
